+++ pagination/common/src/commonMain/kotlin/dev/inmo/micro_utils/pagination/PaginationResult.kt
@@ -1,23 +1,60 @@
 package dev.inmo.micro_utils.pagination
 
-import kotlinx.serialization.Serializable
+import kotlinx.serialization.*
+import kotlin.math.ceil
 
+/**
+ * @param page Current page number
+ * @param size Current page size. It can be greater than size of [results]
+ * @param results Result objects
+ * @param objectsNumber Count of all objects across all pages
+ */
 @Serializable
 data class PaginationResult<T>(
     override val page: Int,
-    val pagesNumber: Int,
+    override val size: Int,
     val results: List<T>,
-    override val size: Int
-) : Pagination
+    val objectsNumber: Long
+) : Pagination {
+    /**
+     * Amount of pages for current pagination
+     */
+    @EncodeDefault
+    val pagesNumber: Int = ceil(objectsNumber / size.toFloat()).toInt()
 
-fun <T> emptyPaginationResult() = PaginationResult<T>(0, 0, emptyList(), 0)
+    constructor(
+        page: Int,
+        results: List<T>,
+        pagesNumber: Int,
+        size: Int
+    ) : this(
+        page,
+        size,
+        results,
+        (pagesNumber * size).toLong()
+    )
+    @Deprecated("Replace with The other order of incoming parameters or objectsCount parameter")
+    constructor(
+        page: Int,
+        pagesNumber: Int,
+        results: List<T>,
+        size: Int
+    ) : this(
+        page,
+        results,
+        pagesNumber,
+        size
+    )
+}
+
+fun <T> emptyPaginationResult() = PaginationResult<T>(0, 0, emptyList(), 0L)
 
 /**
  * @return New [PaginationResult] with [data] without checking of data sizes equality
  */
 fun <I, O> PaginationResult<I>.changeResultsUnchecked(
     data: List<O>
-): PaginationResult<O> = PaginationResult(page, pagesNumber, data, size)
+): PaginationResult<O> = PaginationResult(page, size, data, objectsNumber)
 /**
  * @return New [PaginationResult] with [data] <b>with</b> checking of data sizes equality
  */
@@ -33,12 +70,9 @@ fun <T> List<T>.createPaginationResult(
     commonObjectsNumber: Long
 ) = PaginationResult(
     pagination.page,
-    calculatePagesNumber(
-        commonObjectsNumber,
-        pagination.size
-    ),
+    pagination.size,
     this,
-    pagination.size
+    commonObjectsNumber
 )
 
 fun <T> List<T>.createPaginationResult(
@@ -46,12 +80,9 @@ fun <T> List<T>.createPaginationResult(
     commonObjectsNumber: Long
 ) = PaginationResult(
     calculatePage(firstIndex, size),
-    calculatePagesNumber(
-        commonObjectsNumber,
-        size
-    ),
+    size,
     this,
-    size
+    commonObjectsNumber
 )
 
 fun <T> Pair<Long, List<T>>.createPaginationResult(