+++ repos/common/src/commonMain/kotlin/dev/inmo/micro_utils/repos/mappers/OneToManyKeyValueMappers.kt
@@ -1,7 +1,6 @@
 package dev.inmo.micro_utils.repos.mappers
 
-import dev.inmo.micro_utils.pagination.Pagination
-import dev.inmo.micro_utils.pagination.PaginationResult
+import dev.inmo.micro_utils.pagination.*
 import dev.inmo.micro_utils.repos.*
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
@@ -19,11 +18,8 @@ open class MapperReadKeyValuesRepo<FromKey, FromValue, ToKey, ToValue>(
         pagination,
         reversed
     ).let {
-        PaginationResult(
-            it.page,
-            it.pagesNumber,
-            it.results.map { it.toInnerValue() },
-            it.size
+        it.changeResultsUnchecked(
+            it.results.map { it.toInnerValue() }
         )
     }
 
@@ -34,11 +30,8 @@ open class MapperReadKeyValuesRepo<FromKey, FromValue, ToKey, ToValue>(
         pagination,
         reversed
     ).let {
-        PaginationResult(
-            it.page,
-            it.pagesNumber,
-            it.results.map { it.toInnerKey() },
-            it.size
+        it.changeResultsUnchecked(
+            it.results.map { it.toInnerKey() }
         )
     }
 
@@ -51,11 +44,8 @@ open class MapperReadKeyValuesRepo<FromKey, FromValue, ToKey, ToValue>(
         pagination,
         reversed
     ).let {
-        PaginationResult(
-            it.page,
-            it.pagesNumber,
-            it.results.map { it.toInnerKey() },
-            it.size
+        it.changeResultsUnchecked(
+            it.results.map { it.toInnerKey() }
         )
     }